style: move create div .. for test to main.ts, style fixes
